[TM-1174] move functions to validation helper
